feat(airports): add airports.csv file and remove airports.xlsx file
The airports.csv file is added to the data directory. This file contains a list of airports with their FAA code, name, latitude, and longitude. This file will be used for airport-related operations in the application.

The airports.xlsx file is removed from the data directory. This file was in binary format and is no longer needed. The airports.csv file will be used instead for airport data.